requirements.txt
selenium
pytest
pytest-xdist
webdriver-manager
pillow